+++ packages/photoAlbum/src/index.ts
@@ -1,0 +1,3 @@
+export { default } from '../../../src/components/photoAlbum';
+export { default as PhotoAlbum } from '../../../src/components/photoAlbum';
+export type * from '../../../src/components/photoAlbum';